// src/features/offers/api.types.ts
import { z } from "zod";

export const ApiOffer = z.object({
  id: z.string(),
  created: z.string(), // ISO date string
  updated: z.string(), // ISO date string
  number: z.number(),
  offerText: z.string(),
  showingText: z.string(),
  internalNote: z.string().nullable(),
  deadline: z.string().nullable(),
  recipients: z.array(z.unknown()).optional(),
  recipientsCount: z.number().optional(),
  winnerId: z.string().nullable(),
  state: z.string(),
  organizationId: z.string(),
  organization: z.string().optional(),
  companyId: z.string(),
  company: z.string().optional(),
  residenceId: z.string().optional(),
  propertyId: z.string().optional(),
  residenceAddress: z.string().optional(),
  residencePostalCode: z.number().optional(),
  hasUnreadMessages: z.boolean().optional(),
  unreadMessagesCount: z.number().optional(),
  onlyForStudents: z.boolean().optional(),
});

export type ApiOffer = z.infer<typeof ApiOffer>;

export const ApiOffersPage = z.object({
  facets: z.any(),
  totalResults: z.number(),
  page: z.number(),
  pageSize: z.number(),
  results: z.array(ApiOffer),
});

export type ApiOffersPage = z.infer<typeof ApiOffersPage>;